refactor instrumentation package (#5324)
* refactor instrumentation package

* make linter happy
